Don't mark hud.interface.c for translation.
"make check" is failing in the po/ directory when it can't find hud.interface.c in the srcdir, so it fails if builddir != srcdir. This raises an interesting (to me, at least) question about how to get POTFILES.in to understand where to look for files generated from BUILT_SOURCES. However, there is a shorter path to solving this ticket: since nothing in hud.interface.c needs translation, removing it from POTFILES.in fixes "make check".