distance: drop ignore_character() logic
We currently had logic deep inside the hud's matching algorithm to
ignore a set of characters: _(" _->")

As it is, the characters ' ', '-' and '>' will never been seen by the
algorithm because these characters are used as separators in the earlier
tokenisation process.

'_' is handled now by the HudItem code.  '_F' is turned into just 'F',
for example.  The only way a '_' will end up at this point is if the
application wanted a literal underscore (ie: they wrote '__' in the
label) or if the user explicitly typed it as part of a search term.  In
either of those cases, we should probably be treating it as a normal
character anyway.

That leaves the matter of the possibility that the translators could
introduce additional characters here which, after discussing it, is
probably not reasonable to support just for the sake of it.