HudToken and HudResult: fix up match highlighting
Instead of returning (copies of) strings that we matched, keep track of
the original strings and return them directly.  Then use pointer
comparison to ensure we are highlighting exactly the correct segment.

This avoids accidentally highlighting text that was part of markup
escapes and also ensures that we are highlighting the correct instance
of a string.