Remove window-destroy timeout
The windows were hashed by xid, and the entries in the hash table weren't
removed until five seconds after the window was destroyed.  Thus, restarting an
application within these five seconds and being lucky enough to get the same
xid again would lead to the menu of the old instance being shown.  Activating
menu items doesn't work in that case, because the action group is on the dbus
connection of the new instance.

This was initially introduced to work around a bug in BAMF (lp:718926), which
has since been fixed.