WindowMenuModel: insert action groups on the toplevel menus
Inserting them on the menu items was a workaround for the long-fixed
cloaking-bug in libindicator. It's not supported in gtk 3.10 anymore, because
the menu tracker in gtk doesn't allow action groups on menu items. That's for
good reasons: it's more efficient and consumers of that api should not modify
the created menu hierarchy.